# docs/swagger.json
{
	"components": {
...
		"schemas": {
...
			"IProduct": {
				"properties": {
					"title": {
						"type": "string"
					},
					"in_stock": {
						"type": "boolean"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"discount": {
						"type": "number",
						"format": "double"
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"type": {
						"type": "string",
						"enum": [
							"product",
							"service"
						]
					},
					"category": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"short_description": {
						"type": "string"
					},
					"images": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"quantity_sold": {
						"type": "number",
						"format": "double"
					},
					"quantity_viewed": {
						"type": "number",
						"format": "double"
					},
					"quantity_in_cart": {
						"type": "number",
						"format": "double"
					},
					"quantity_in_wishlist": {
						"type": "number",
						"format": "double"
					},
					"quantity_in_compare": {
						"type": "number",
						"format": "double"
					},
					"brand": {
						"type": "string"
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"uploaded_by": {
						"type": "string"
					}
				},
				"required": [
					"price"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "peddle",
		"version": "1.0.0",
		"description": "a basic express app writen in typescript",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Abdulraheem A omeiza"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/products/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"anyOf": [
												{
													"$ref": "#/components/schemas/IProduct"
												},
												{}
											]
										},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"409": {
						"description": "email already in use"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"short_description": {
										"type": "string"
									},
									"brand": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"tags": {
										"type": "string"
									},
									"discount": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			}
		},
		"/products/update": {
			"patch": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Not allowed"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"title": {
										"type": "string"
									},
									"description": {
										"type": "string"
									},
									"short_description": {
										"type": "string"
									},
									"price": {
										"type": "string"
									},
									"brand": {
										"type": "string"
									},
									"quantity": {
										"type": "string"
									},
									"category": {
										"type": "string"
									},
									"tags": {
										"type": "string"
									},
									"discount": {
										"type": "string"
									}
								}
							}
						}
					}
				}
			}
		},
		"/products/product": {
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Not allowed"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"productId": {
										"type": "string"
									}
								},
								"required": [
									"productId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/products/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Products not found"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": []
			}
		},
		"/products/get/{id}": {
			"get": {
				"operationId": "GetProduct",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"additionalProperties": false,
											"type": "object"
										},
										"status": {
											"type": "number",
											"format": "double"
										},
										"success": {
											"type": "boolean",
											"enum": [
												false
											],
											"nullable": false
										}
									},
									"required": [
										"message",
										"status",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Product not found"
					}
				},
				"tags": [
					"Products"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}